order-service/requirements.txt
flask==3.1.1
flasgger==0.9.7.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1